+++ src/LichessSharp/Serialization/LichessJsonContext.cs
@@ -82,6 +82,8 @@ namespace LichessSharp.Serialization;
 [JsonSerializable(typeof(List<Streamer>))]
 [JsonSerializable(typeof(StreamInfo))]
 [JsonSerializable(typeof(StreamerInfo))]
+[JsonSerializable(typeof(UserStreamer))]
+[JsonSerializable(typeof(StreamChannel))]
 [JsonSerializable(typeof(Timeline))]
 [JsonSerializable(typeof(TimelineEntry))]
 [JsonSerializable(typeof(TimelineData))]